Also draw b/w 640x384 and 3-color 400x300 bitmaps on GDEW075Z09
Since there are no example 3-color 640x384 bitmaps, I have modified the example so that 1-color 640x384 bitmaps and 3-color 400x300 bitmaps are drawn if the GDEW075Z09 display is selected.
For the 400x300 bitmaps, I have modified the method (which originally worked only if the display had the exact same size as the bitmaps) so that it works as expected also on bigger displays, such as the GDEW075Z09. In that case they are centered.

Only tested on WaveShare 7.5-inch bwr display (and ESP8266).